refactor(ui): ui package bundling, closes leather-io/issues#351
