+++ .github/actions/provision/action.yml
@@ -6,7 +6,7 @@ runs:
     - name: Install pnpm
       uses: pnpm/action-setup@v4
       with:
-        version: '9.6.0'
+        version: '9.11.0'
 
     - name: Set up node
       uses: actions/setup-node@v3
