+++ packages/ui/src/components/approver/animations/approver-animation.web.tsx
@@ -1,8 +1,8 @@
 import { motion, stagger, useAnimate } from 'framer-motion';
 import { css } from 'leather-styles/css';
 
-import { HasChildren } from '../../../utils/has-children';
-import { useOnMount } from '../../../utils/use-on-mount';
+import { HasChildren } from '../../../utils/has-children.shared';
+import { useOnMount } from '../../../utils/use-on-mount.shared';
 
 const animationSelector = '& > *:not(.skip-animation)';
