+++ packages/ui/src/components/toast/toast.layout.web.tsx
@@ -1,6 +1,6 @@
 import { Box, HStack, styled } from 'leather-styles/jsx';
 
-import { getIconVariant } from './toast.utils';
+import { getIconVariant } from './toast.utils.web';
 import { type ToastProps } from './toast.web';
 
 export function ToastLayout({ message, variant }: ToastProps) {
