+++ packages/ui/src/utils/use-register-children.shared.ts
